:new: #2361 【企业微信】增加客户朋友圈相关的部分接口

+++ weixin-java-cp/src/main/java/me/chanjar/weixin/cp/bean/external/WxCpUserExternalTagGroupInfo.java
@@ -35,7 +35,7 @@ public class WxCpUserExternalTagGroupInfo extends WxCpBaseResp {
     private Long createTime;
 
     @SerializedName("order")
-    private Integer order;
+    private Long order;
 
     @SerializedName("deleted")
     private Boolean deleted;
@@ -67,7 +67,7 @@ public class WxCpUserExternalTagGroupInfo extends WxCpBaseResp {
     private Long createTime;
 
     @SerializedName("order")
-    private Integer order;
+    private Long order;
 
     @SerializedName("deleted")
     private Boolean deleted;
